cleaning up querySelectors whose value is not being read in their files
