+++ src/domUpdates/domUpdates.js
@@ -5,8 +5,18 @@ import { getTripDetailsForTraveler, calculateTotalAmountSpent, categorizeTrips }
 //querySelectors
 const totalAmountSpentElement = document.querySelector('.sub-container4 .card-DOMUpdates')
 const pastTripsElement = document.querySelector('.sub-container2 .card-DOMUpdates')
-
-
+const loginButton = document.getElementById('loginButton');
+const bookTripButton = document.querySelector('.nav-book-button');
+const pendingTripsText = document.querySelector('.pending-card-DOMUpdates');
+const upcomingTripsText = document.querySelector('.upcoming-card-DOMUpdates');
+const upcomingTripsHeader = document.querySelector('.sub-container1 .card-titles');
+const pastTripsHeader = document.querySelector('.sub-container2 .card-titles');
+const pendingTripsHeader = document.querySelector('.sub-container3 .card-titles');
+const amountSpentHeader = document.querySelector('.sub-container4 .card-titles');
+const bookingSection = document.querySelector('.booking-section');
+bookTripButton.style.display = 'none';
+const tripMessage = document.querySelector('.trip-message');
+const tripMessageSubContainer = document.querySelector('.sub-container-text');
 
 const updateTotalAmountSpent = (trips = [], destinations = [], currentYear) => {
   const approvedTrips = trips.filter(trip => {
@@ -52,5 +62,26 @@ const hideLoginForm = () => {
   loginForm.classList.add("login-form-hidden");
 };
 
-export { updateTotalAmountSpent, updatePastTrips, showLoginForm, hideLoginForm }
+const handleLogout = () => {
+  const usernameInput = document.getElementById('username');
+  const passwordInput = document.getElementById('password');
+  usernameInput.value = '';
+  passwordInput.value = '';
+  upcomingTripsHeader.textContent = 'Explore the World Stress-Free!';
+  upcomingTripsText.textContent = 'Discover over 50 dream destinations hassle-free with our expert planning. Let us handle every detail while you focus on the adventure!';
+  pastTripsHeader.textContent = 'Unlock Your Wanderlust with Ease!';
+  pastTripsElement.textContent = 'With a nominal 10% agent fee, enjoy the convenience of seamless travel planning. Leave the logistics to us and embrace the joy of exploration!';
+  pendingTripsHeader.textContent = 'Your Journey, Our Expertise';
+  pendingTripsText.innerHTML = "From booking to boarding, we've got you covered. Trust our seasoned agents to curate your perfect getaway, ensuring a memorable experience every step of the way!";
+  amountSpentHeader.textContent = "Let's get started!";
+  totalAmountSpentElement.textContent = 'Begin your journey by logging in. Your adventure awaits just a click away!';
+  bookTripButton.style.display = 'none';
+  loginButton.innerText = "Login";
+  bookingSection.style.display = 'none';
+  tripMessage.textContent = 'Welcome, adventurer!';
+  tripMessageSubContainer.lastElementChild.textContent = 'Let us book your travel experience!';
+};
+
+
+export { updateTotalAmountSpent, updatePastTrips, showLoginForm, hideLoginForm, handleLogout }
